Remove the original edit campaign
