+++ app.py
@@ -68,8 +68,8 @@ nav = st.navigation(
             title="Edit Campaign Template",
         ),
         st.Page(
-            "pages/va/edit_campaign_v2.py",
-            title="Edit Campaign Template V2 Beta",
+            "pages/va/edit_campaign.py",
+            title="Edit Campaign Template",
         ),
         st.Page(
             "pages/va/revert_campaign_template.py",
